Some fixes

+++ utils/vmpi/vmpi_filesystem_master.cpp
@@ -5,6 +5,8 @@
 //=============================================================================//
 
 #include <winsock2.h>
+
+#include "fasttimer.h"
 #include "vmpi_filesystem_internal.h"
 #include "zlib.h"
 #include "vstdlib/random.h"
@@ -707,7 +709,7 @@ void CMasterMulticastThread::TCP_SendNextChunk( CMulticastFile *pFile, CClientFi
 		sizeof( cPacket ),
 		sizeof( pFile->m_Info ),
 		sizeof( m_iCurActiveChunk ),
-		strlen( pFile->GetFilename() ) + 1,
+		(int)strlen( pFile->GetFilename() ) + 1,
 		iEndByte - iStartByte
 	};
